[Widgets] refactoring and limits table vertical size in windows
